task4/requirements.txt
pandas>=2.0
numpy>=1.25
matplotlib>=3.8
pyyaml>=6.0
python-dateutil>=2.9
streamlit>=1.30
tabulate>=0.9
pyarrow>=12.0

